+++ src/guiengine/widgets/dynamic_ribbon_widget.cpp
@@ -46,7 +46,7 @@ DynamicRibbonWidget::DynamicRibbonWidget(const bool combo, const bool multi_row)
     m_scrolling_enabled    = true;
 
     // by default, set all players to have no selection in this ribbon
-    for (int n=0; n<MAX_PLAYER_COUNT; n++)
+    for (unsigned int n=0; n<MAX_PLAYER_COUNT; n++)
     {
         m_selected_item[n] = -1;
     }
@@ -305,7 +305,7 @@ void DynamicRibbonWidget::buildInternalStructure()
     //printf("****DynamicRibbonWidget::buildInternalStructure()****\n");
 
     // ---- Clean-up what was previously there
-    for (int i=0; i<m_children.size(); i++)
+    for (unsigned int i=0; i<m_children.size(); i++)
     {
         IGUIElement* elem = m_children[i].m_element;
         if (elem != NULL && m_children[i].m_type == WTYPE_RIBBON)
@@ -432,7 +432,7 @@ void DynamicRibbonWidget::buildInternalStructure()
     {
         // debug checks
         int childrenCount = 0;
-        for (int n=0; n<m_rows.size(); n++)
+        for (unsigned int n=0; n<m_rows.size(); n++)
         {
             childrenCount += m_rows[n].m_children.size();
         }
@@ -713,7 +713,7 @@ void DynamicRibbonWidget::onRibbonWidgetFocus(RibbonWidget* emitter, const int p
 {
     if (m_deactivated) return;
 
-    if (emitter->m_selection[playerID] >= emitter->m_children.size())
+    if (emitter->m_selection[playerID] >= (int)emitter->m_children.size())
     {
         emitter->m_selection[playerID] = emitter->m_children.size()-1;
     }
@@ -759,7 +759,7 @@ void DynamicRibbonWidget::scroll(const int x_delta)
     // update selection markers in child ribbon
     if (m_combo)
     {
-        for (int n=0; n<MAX_PLAYER_COUNT; n++)
+        for (unsigned int n=0; n<MAX_PLAYER_COUNT; n++)
         {
             RibbonWidget* ribbon = m_rows.get(0); // there is a single row when we can select items
             int id = m_selected_item[n] - m_scroll_offset;
@@ -775,7 +775,7 @@ void DynamicRibbonWidget::scroll(const int x_delta)
  used to ensure that all children ribbons always select the same column */
 void DynamicRibbonWidget::propagateSelection()
 {
-    for (int p=0; p<MAX_PLAYER_COUNT; p++)
+    for (unsigned int p=0; p<MAX_PLAYER_COUNT; p++)
     {
         // find selection in current ribbon
         RibbonWidget* selected_ribbon = (RibbonWidget*)getSelectedRibbon(p);
@@ -1025,7 +1025,7 @@ bool DynamicRibbonWidget::findItemInRows(const char* name, int* p_row, int* p_id
     int row = -1;
     int id = -1;
 
-    for (int r=0; r<m_rows.size(); r++)
+    for (unsigned int r=0; r<m_rows.size(); r++)
     {
         id = m_rows[r].findItemNamed(name);
         if (id > -1)
